requirements.txt
requests
tus.py